requirements.txt
fastapi[all]>=0.100.0       # Includes uvicorn, pydantic, etc.
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.27.0           # Driver for PostgreSQL (use 'aiosqlite' for SQLite)
alembic>=1.9.0
pydantic-settings>=2.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.5   # For form data (OAuth2PasswordRequestForm)
email-validator>=1.3.0    # For email validation in Pydantic schemas
uuid>=1.30                # For UUID generation